pyproject.toml
[project]
name = "ai-lab-investment"
version = "0.0.1"
description = "This is a template repository for research projects that use Python and Quarto."
authors = [
    { name = "Vincent Grégoire", email = "vincent.gregoire@gmail.com" },
]
readme = "README.md"
keywords = ['python']
requires-python = ">=3.12,<4.0"
classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = ["hydra-core>=1.3.2", "python-dotenv>=1.1.1"]

[project.urls]
Homepage = "https://vgreg.github.io/ai-lab-investment/"
Repository = "https://github.com/vgreg/ai-lab-investment"
Documentation = "https://vgreg.github.io/ai-lab-investment/"

[dependency-groups]
dev = [
    "pytest>=7.2.0",
    "pre-commit>=2.20.0",
    "ty>=0.0.1a16",
    "pytest-cov>=4.0.0",
    "ruff>=0.11.5",
    "mkdocs>=1.4.2",
    "mkdocs-material>=8.5.10",
    "mkdocstrings[python]>=0.26.1",
]

[build-system]

requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.ty.environment]


python = "./.venv"
python-version = "3.13"
[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py313"
line-length = 88
fix = true

[tool.ruff.lint]
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E",
    "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
    # NumPy-specific rules,
    "NPY",
    # pandas-vet
    "PD",
]
ignore = []

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.ruff.format]


preview = true
[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
source = ["src"]

branch = true
